Added manager container for low-level APIs, refactored scala virtual machine to expose managers via lowlevel field
