+++ scala-debugger-api/src/it/scala/org/senkbeil/debugger/api/lowlevel/methods/MethodEntryManagerIntegrationSpec.scala
@@ -35,18 +35,20 @@ class MethodEntryManagerIntegrationSpec extends FunSpec with Matchers
       val reachedMethodBeforeFirstLine = new AtomicBoolean(false)
 
       withVirtualMachine(testClass, suspend = false) { (v, s) =>
+        import s.lowlevel._
+
         // Set up the method entry event
-        s.methodEntryManager.setMethodEntry(
+        methodEntryManager.setMethodEntry(
           expectedClassName,
           expectedMethodName
         )
 
         // First line in test method
-        s.breakpointManager.setLineBreakpoint(testFile, 26)
+        breakpointManager.setLineBreakpoint(testFile, 26)
 
         // Listen for breakpoint on first line of method, checking if this
         // breakpoint is hit before or after the method entry event
-        s.eventManager.addResumingEventHandler(BreakpointEventType, e => {
+        eventManager.addResumingEventHandler(BreakpointEventType, e => {
           val breakpointEvent = e.asInstanceOf[BreakpointEvent]
           val location = breakpointEvent.location()
           val fileName = location.sourcePath()
@@ -59,7 +61,7 @@ class MethodEntryManagerIntegrationSpec extends FunSpec with Matchers
         })
 
         // Listen for method entry events for the specific method
-        s.eventManager.addResumingEventHandler(MethodEntryEventType, e => {
+        eventManager.addResumingEventHandler(MethodEntryEventType, e => {
           val methodEntryEvent = e.asInstanceOf[MethodEntryEvent]
           val method = methodEntryEvent.method()
           val className = method.declaringType().name()
